Removing Firefox 3.6 support. It has reached EOL. Details below:
Mozilla killed off 3.6 officially and it's now two days dead:
  https://blog.mozilla.org/futurereleases/2012/03/23/upcoming-firefox-support-changes/

Firefox ESR is their replacement by the way, it's a Firefox that's updated ~once a year:
  https://wiki.mozilla.org/Enterprise/Firefox/ExtendedSupport:Proposal
  http://www.mozilla.org/en-US/firefox/organizations/

Usage data is trending down, and it's at ~2.68% worldwide.
  http://gs.statcounter.com/#browser_version-ww-daily-20120416-20120416-bar

Oh, look at that, IE7 has the same # of users! Guess who is next to sunset support for? :)